requirements.txt
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-markdownify==0.8.0
django-markdownx==2.0.28
feedparser==5.2.1
idna==2.7
Markdown==3.0.1
Pillow==5.3.0
pkg-resources==0.0.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
webencodings==0.5.1
